requirements.txt
earthkit-data
earthkit-maps
earthkit-meteo
xarray
netCDF4
scipy
matplotlib
cartopy
ecmwf-opendata
cdsapi
pandas
scikit-learn
xgboost
geopandas
rasterio
matplotlib
folium
mapclassify
fastapi
uvicorn
geopy
duckdb
